PIC-TEXT-TO-STR alpha 0.0.1
The bot is able to recognize the text on the image.

At this stage, the bot is able to qualitatively recognize only clearly distinguished text on the image.

Language switch.

There is a configuration file for more convenient use of the bot. It is created automatically if suddenly it is missing.

Auto-creation of the user's image folder.  Checking the presence of the user's folder.

Auto-delete user images.